backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyserial==3.5
python-multipart==0.0.6
Pillow==10.1.0
opencv-python-headless==4.8.1.78
# NumPy 1.26+ supports Python 3.12 and keeps vpype deps happy
numpy>=1.26,<2.0
scipy>=1.11.0
scikit-image>=0.22.0
python-dotenv==1.0.0
websockets==12.0
aiofiles==23.2.1
debugpy==1.8.0
PyYAML==6.0.1
vpype[all]==1.15.0
# Adds gcode export command for vpype
vpype-gcode>=0.10.0
svgpathtools==1.6.1
